pyproject.toml
[project]
name = "carsshop"
version = "0.1.0"
description = ""
authors = [
    {name = "aziret"}
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "django (>=5.2.5,<6.0.0)",
    "aiogram (==2.25.1)",
    "django-jazzmin (>=3.0.1,<4.0.0)",
    "pillow (>=11.3.0,<12.0.0)",
    "moviepy (>=2.2.1,<3.0.0)",
    "python-decouple (>=3.8,<4.0)",
    "dj-database-url (>=3.0.1,<4.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
